requirements.txt
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
click>=8.0.0
tiktoken>=0.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6